fix: clean up pointer casts and size formats
Align legacy C casts and printf formats with pointer-sized types to reduce compiler warnings without changing behavior.

Warning cleanup:
  - Use unsigned char for Cineon magic-number byte inspection in cininfo.c and cast orientation strings at the printf boundary
  - Print RibGetBufferLength size_t values with %zu in strmtest.c
  - Route fake RI handles through size_t before object-pointer casts in ribnop.c and write.c
  - Return integer zero from RibSetHintTable's comma expression in ribrdr.h

+++ src/ributil/ribnop.c
@@ -193,7 +193,7 @@ RtLightHandle RiNopAreaLightSourceV( RtToken name,
    (void)name; (void)n; (void)tokens; (void)parms;
    
    /* Fake it. */
-   return (RtLightHandle)LastLightHandle++;
+   return (RtLightHandle)((size_t)LastLightHandle++);
 }
 
 
@@ -490,7 +490,7 @@ RtLightHandle RiNopLightSourceV( RtToken name,
    (void)name; (void)n; (void)tokens; (void)parms;
 
    /* Fake it. */
-   return (RtLightHandle)LastLightHandle++;
+   return (RtLightHandle)((size_t)LastLightHandle++);
 }
 
 
@@ -600,7 +600,7 @@ RtVoid RiNopNuPatchV( RtInt nu, RtInt uorder, RtFloat uknot[],
 RtObjectHandle RiNopObjectBegin( void )
 {
    /* Fake it. */
-   return (RtObjectHandle)LastObjectHandle++;
+   return (RtObjectHandle)((size_t)LastObjectHandle++);
 }
